# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Pokemon Cards API - OpenAPI 3.0",
    "description": "This is a sample Pokemon Cards API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "pedro.jose.bidoggia@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "http://localhost:8000/api"
    }
  ],
...
  "paths": {
...
    "/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Get user token",
        "description": "",
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "auth": [
              "write:pokemons",
              "read:pokemons"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "theUser"
          },
          "email": {
            "type": "string",
            "example": "john@email.com"
          },
          "password": {
            "type": "string",
            "example": "12345"
          }
        },
        "xml": {
          "name": "user"
        }
      },
...
    },
...
    "securitySchemes": {
      "auth": {
        "type": "apiKey",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://localhost:8000",
            "scopes": {
              "write:pokemons": "modify pokemons in your account",
              "read:pokemons": "read your pokemons"
            }
          }
        }
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}